Fix false positives in test_domain_projection
This simplistic test was checking that certain values were present in the proj string for the domain projection. However, some values such as XCENT appeared only coincidentally, since they shared the same value as MOAD_CEN_LAT. Additionally, values from the domain file are being cast to `float` in the projection, causing the tests to succeed when values are all rounded down, but fail when a value is rounded up, which is the case when using the aust-test domain.

This improved test checks that values are assigned using the correct projection attribute, and that values are cast to float.