Normalise layer dimensions
This change ensures that all layers share the dimensions ("time", "vertical", "y", "x"). While the "vertical" layer is a single-value in all layers, this may be expanded later.

For the "time" dimension, some datasets such as GFAS will have emissions for each day in the period, while most will be a single set of average emissions that are valid across the entire period. We simply duplicate the data for the layers without time step data to fill out the correct number of entries in the time dimension.

On one hand, duplicating the data doesn't add anything new and is likely to increase the final file size. On the other hand, this gives us a normalised layer format so that operations on one layer can be performed on every layer without special handling.