Rename "OCH4" variable names in output
The new names are more consistent with the rest of the file and CF Conventions, and the "och4_sector" prefix on sector layers makes them easily distinguishable from the total. This also meant replacing the use of "layer" in much of our code with "sector".

The OCH4_TOTAL variable has been left intact temporarily while we refactor downstream consumers.