Replace lat_bounds and lon_bounds with x_bounds and y_bounds in output
Including lat_bounds and lon_bounds this way is actually not valid in CF Conventions, which dictates that bounds can be added only for coordinate variables. Now that the "x" and "y" coordinates store grid projection coordinates, we can use "x_bounds" and "y_bounds", plus the details of the projection, to allow clients to accurately re-create the grid cell boundaries.

This also takes up dramatically less space on disk.