Add domain "slug"
Add a short, URL-safe "slug" to the domain metadata. The purpose of this string is to be used in things like front-end ids or URLs where the domain name might be too long.

Ideally this would be an attribute in the domain file itself, but until that is addressed it can be added here so it will be available downstream.

Resolves #40.