Add utilities for generating grid cell names
Grid cell names are short, unique identifiers which can be used downstream for things like the web front-end, so URLs for a single cell can be constructed. Sharing this unique id across the back-end and front-end will make it easier to identify cells in the data that have become of interest due to queries on the front-end.