Add "cell_name" to prior output
This variable stores a string for each cell in the grid which corresponds to its "display name" visible on the front-end. Generating these here and passing them downstream ensures we don't need to duplicate code for generating these names across multiple projects and languages, and makes it easier to identify cells in the data based on findings in the front-end.