Remove @property decorators from Grid due to poor performance
Using @property to define the derived attributes of the Grid made code that felt very easy to read, and meant that derived properties would always be up to date if dependent attributes were changed.

However, when the raster_remap method (which does calculations over thousands of cells) was refactored to use Grid properties, the execution time more than doubled from 4 seconds to 11 seconds. This doesn't seem worth it just for syntactic sugar, so this commits updates Grid to use simple attribute setting in __init__ for derived properties.