Remove intermediate reprojections which are no longer used
Both of these intermediate reprojections were problematic, and led to subtle bugs that have now been fixed. Although it would be possible to store the output of these reprojections for re-use, this hasn't been actioned, so this step should just be removed in the meantime, to stop wasting resources on reprocessing that isn't used.