Disable progress bar for GFAS download
Similar to the progress bar which was removed from the livestock layer, this progress bar is problematic in logged environments such as GHA or AWS production runs where progress ends up being written over many lines. Although this is manageable via env variables, the GFAS download doesn't take long enough to justify a progress bar and the complexity that managing it would bring.