Rename create_prior_domain "landuse" argument "inventory-geotiff"
The way the landuse file is being used it actually quite generic. The GeoTIFF is loaded, all non-zero values are converted to 1, and the raster is regridded to the domain grid. By naming this a more generic name and explaining what's expected, users can safely supply their own inventory GeoTIFF.