Add additional snapshot tests to prior result
The current snapshot test runs the entire prior and compares the results to known good values by examining the mean value of entire sectors. This adds similar tests, but includes max values to detect other fluctuations, as well as an x band and y band test to guard against shifts in geometry.