Simplify DOMAIN configuration by allowing a URL
Users can now reference their domain of interest (and inventory domain) by simply providing an absolute or relative file path, or a URL where the domain can be downloaded.

If a URL is provided, omDownloadInputs.py will fetch the file if not already present.

Note that specifying a DOMAIN_NAME and DOMAIN_VERSION is no longer supported.