Add methods for working with UNFCCC sector category data
Sectors in the prior are annotated with a UNFCCC sector code, but the Australian GHG inventory API doesn't use UNFCCC codes. The AU inventory uses the full names of UNFCCC sectors (some of which are not exactly the same as names in the CRT tables).

These helpers allow us to turn a CSV of UNFCCC codes and level names into a searchable resource so we can match API data with sector codes.