Add methods for working with Australian GHG inventory data
The Australian National Greenhouse Accounts makes our annual emissions inventory available by UNFCCC sector in the "Paris Agreement" inventory, which we've transformed into a CSV. These helper functions allow us to load the CSV inventory into a searchable data structure that can be used to get the sum of emissions within a certain category for a designated time period.

This will be useful for establishing a clear emission total for each prior sector for any period within the available data.