Add option to parse DataSource on load
Parsed DataSources are particularly useful for data sources which are shared across multiple layers, which have an expensive parsing or preprocessing step. The resource will be fetched and parsed once, and the result cached by the DataManager.

Several examples have been implemented of CSV data sources, where the parse() method will interpret the raw file data by reading it in as a pandas DataFrame. This is provided directly by the DataManager without the calling code having to implement file parsing.