Add SafeguardFacility class and helper methods
SafeguardFacility represents a single physical location, and the emissions reported for it. While sometimes a single facility will appear over multiple rows in the SGM, for example when the controlling business has changed in the middle of a reporting period, we want to represent the facility only once, with all its emissions over the period.