Refactor run_prior into create_prior
This name is more descriptive of what the method is doing. Additionally, verifying and writing output is moved out of the create_prior method and is now the responsibility of scripts which might use it. This makes testing simpler, and gives the user more freedom to choose an output format in their own script.