Create a single snapshots for test_009_prior_emissions_ds
Originally, this test makes several assertions against snapshots of expected scientific results. However, when changes are made to the codebase that cause expected changes, this test had to be run multiple times (very slowly) to discover and fix the changes to each different comparison.

This change gathers all the calculations for the result into a single large snapshot, which can be checked in a single assertion. When values change, the entire new snapshot can be used to replace the old.