Properly account for CRS curvature in regrid_any
The new approach uses the min and max lon/lat values in the target grid to construct a bounding box with four corners in the source grid. This means possibly more source grid cells are examined but should guarantee that no source grid cells are missed due to alignment/curvature differences.