Rework remap_raster to clip the input to the target grid bounding box
In the previous implementation, remap_raster aggregated values for the entire input dataset, ignoring the spatial extent of the target grid. This is very inefficient in cases where the target grid is much smaller than the input raster (ie the au-test domain).

This change constructs a bounding box that covers the target grid, and converts it to coordinates in the input field. This is used to subset the input field before aggregating input values.

On my machine, this halves the time it takes to run the test suite.